OVERHAUL OF GAME2.LUA ("VOCAB")
This commit is the initial overhaul of game2.lua. The drastic changes to this file were mainly for visual purposes of the game.

The main changes to the code structure was the addition of coordinate varaibles (in order to control positions of images exactly), as well as some supporting methods. However, in order to get this working, I kept the method structure from the last developer: this may eventually need to be fixed, as the methods jump around a lot and are a bit vague.

Testing and cleaning will need to be done on this commited code (representing about 7 hours of total work), but I needed this shown and saved for the moment.

Keep Calm, and Code On...