1.3.3 (15) Commit Switched the Game4 evaluateAnswer() method to convert the chosenAns and the correctAns to numbers, in order to do a (hopefully) successful comparision between them.
This seems like quite a stupid workaround, and I agree, but according to the logs of the android console, Andriod Corona/Lua is not seeing the comparision of chosenAns == correctAns (where both are equal to the string "A") as true. As such I am attempting to switch them to numbers in order to remove any possible confusion.